Moving SqlServer.Design and Sqlite.Design classes into Internal namespaces.
