+++ src/Microsoft.EntityFrameworkCore.SqlServer.Design/Internal/SqlServerTableSelectionSetExtensions.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Annotations;
 
-namespace Microsoft.EntityFrameworkCore.Scaffolding
+namespace Microsoft.EntityFrameworkCore.Scaffolding.Internal
 {
     internal static class SqlServerTableSelectionSetExtensions
     {
