+++ src/Microsoft.EntityFrameworkCore.SqlServer.Design/Metadata/Internal/SqlServerDatabaseModelAnnotationNames.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.EntityFrameworkCore.Scaffolding.Metadata
+namespace Microsoft.EntityFrameworkCore.Scaffolding.Metadata.Internal
 {
     public static class SqlServerDatabaseModelAnnotationNames
     {
