+++ src/Microsoft.EntityFrameworkCore.SqlServer.Design/Metadata/Internal/SqlServerDatabaseModelAnnotations.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Utilities;
 
-namespace Microsoft.EntityFrameworkCore.Scaffolding.Metadata
+namespace Microsoft.EntityFrameworkCore.Scaffolding.Metadata.Internal
 {
     public class SqlServerDatabaseModelAnnotations
     {
