+++ src/Microsoft.EntityFrameworkCore.SqlServer.Design/Metadata/Internal/SqlServerDatabaseModelExtensions.cs
@@ -3,7 +3,7 @@
 
 using JetBrains.Annotations;
 
-namespace Microsoft.EntityFrameworkCore.Scaffolding.Metadata
+namespace Microsoft.EntityFrameworkCore.Scaffolding.Metadata.Internal
 {
     public static class SqlServerDatabaseModelExtensions
     {
