+++ src/Microsoft.EntityFrameworkCore.Sqlite.Design/Internal/SqliteScaffoldingModelFactory.cs
@@ -8,7 +8,7 @@ using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Storage;
 using Microsoft.Extensions.Logging;
 
-namespace Microsoft.EntityFrameworkCore.Scaffolding
+namespace Microsoft.EntityFrameworkCore.Scaffolding.Internal
 {
     public class SqliteScaffoldingModelFactory : RelationalScaffoldingModelFactory
     {
