+++ src/Microsoft.EntityFrameworkCore.Sqlite.Design/Internal/SqliteTableSelectionSetExtensions.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Linq;
 
-namespace Microsoft.EntityFrameworkCore.Scaffolding
+namespace Microsoft.EntityFrameworkCore.Scaffolding.Internal
 {
     internal static class SqliteTableSelectionSetExtensions
     {
