+++ test/Microsoft.EntityFrameworkCore.SqlServer.Design.Tests/ApiConsistencyTest.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.Reflection;
-using Microsoft.EntityFrameworkCore.Scaffolding;
+using Microsoft.EntityFrameworkCore.Scaffolding.Internal;
 using Microsoft.EntityFrameworkCore.Tests;
 
 namespace Microsoft.EntityFrameworkCore.SqlServer.Design.Tests
