+++ test/Microsoft.EntityFrameworkCore.SqlServer.Design.Tests/SqlServerTableSelectionSetExtensionsTests.cs
@@ -4,6 +4,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore.Scaffolding;
+using Microsoft.EntityFrameworkCore.Scaffolding.Internal;
 using Xunit;
 
 namespace Microsoft.EntityFrameworkCore.SqlServer.Design
