+++ test/Microsoft.EntityFrameworkCore.Sqlite.Design.FunctionalTests/SqliteDatabaseModelFactoryTest.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Scaffolding;
+using Microsoft.EntityFrameworkCore.Scaffolding.Internal;
 using Microsoft.EntityFrameworkCore.Scaffolding.Metadata;
 using Microsoft.EntityFrameworkCore.Sqlite.FunctionalTests;
 using Microsoft.Extensions.DependencyInjection;
