+++ test/Microsoft.EntityFrameworkCore.Sqlite.Design.Tests/SqliteTableSelectionSetExtensionsTests.cs
@@ -4,6 +4,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore.Scaffolding;
+using Microsoft.EntityFrameworkCore.Scaffolding.Internal;
 using Xunit;
 
 namespace Microsoft.EntityFrameworkCore.Sqlite.Design
